@@ -22,7 +22,6 @@
     "query-string": "^6.12.1",
     "randomstring": "^1.1.5",
     "redis": "^3.0.2",
-    "util": "^0.12.3",
     "winston": "^3.2.1"
   },
   "dependenciesComments": {
@@ -38,7 +37,6 @@
     "run-container": "^1.2.1",
     "supertest": "^4.0.2",
     "ts-jest": "^24.0.2",
-    "ts-node": "^8.3.0",
-    "typescript": "^3.7.2"
+    "typescript": "^3.8.3"
   }
 }
