@@ -1,85 +1,107 @@
 import { runSimple } from 'run-container'
-import redis from 'redis';
+import redis, { RedisClient } from 'redis';
 import mockyeah from 'mockyeah';
 import retryStrategy from 'node-redis-retry-strategy';
 import { processMsg } from '../processMsg';
 import { ReqRes } from '../reqres';
 import { relayService } from '../relayService';
-const request = require('supertest');
+import request from 'supertest';
+import { AssertionError } from 'assert';
 
-const msg: ReqRes = { 
-  id: 'myId', 
-  startTime: 0, 
-  duration: 2, 
-  method: 'patch', 
-  url: {url: '/test-url', query: {k1: 'v1', k2: 'v2'}},
-  reqBody: {key1: 'abc', key2: '123'},
-  statusCode: 3, 
-  statusMessage: 'myMsg' };
+const msg: ReqRes = {
+  id: 'myId',
+  startTime: 0,
+  duration: 2,
+  method: 'patch',
+  url: { url: '/test-url', query: { k1: 'v1', k2: 'v2' } },
+  reqBody: { key1: 'abc', key2: '123' },
+  statusCode: 3,
+  statusMessage: 'myMsg'
+};
 const rHost = '127.0.0.1';
 const rPort = 6379;
-const channel = 'test-channel';
-let publisher: redis.RedisClient;
-let subscriber: redis.RedisClient;
+let publisher: RedisClient;
+let subscriber1: RedisClient, subscriber2: RedisClient, objectEqualitySubscriber: RedisClient;
 let redisContainer;
 let relay;
-
-const thePromise = () => {
-  return new Promise((resolve, reject) => {
-    subscriber.on('message', async (channel, message) => {
-      const subObj: ReqRes = JSON.parse(message);
-      console.log(`message received: ${JSON.stringify(subObj)}`);
-      resolve(subObj);
-    });
-  });
-}
 
 beforeAll(async () => {
   // Start Redis
   redisContainer = await runSimple({
     image: 'redis',
     name: 'redis-in-jest',
-    ports: { '6379' : '6379' }
+    ports: { '6379': '6379' }
   });
 
   // Sit for 3s
   await new Promise(resolve => setTimeout(resolve, 3000));
 
-  publisher = redis.createClient({host: rHost, port: rPort, retry_strategy: retryStrategy });
-  subscriber = redis.createClient({host: rHost, port: rPort});
-  subscriber.subscribe(channel);
-  relay = relayService({targetUrl: 'http://localhost:4001', client: publisher, topic: channel});
+  publisher = redis.createClient({ host: rHost, port: rPort, retry_strategy: retryStrategy });
+  subscriber1 = redis.createClient({ host: rHost, port: rPort, retry_strategy: retryStrategy });
+  subscriber2 = redis.createClient({ host: rHost, port: rPort, retry_strategy: retryStrategy });
+  objectEqualitySubscriber = redis.createClient({ host: rHost, port: rPort, retry_strategy: retryStrategy });
+  relay = relayService({ targetUrl: 'http://localhost:4001', client: publisher, topic: 'relay-channel' });
 
-  return new Promise((resolve) => setTimeout(() => resolve(), 1000));
+  subscriber1.on('message', (channel, message) => {
+    console.log(`subscriber1 listening ${channel}: ${message}`);
+  });
+
+  subscriber2.on('message', (channel, message) => {
+    console.log(`subscriber2 listening ${channel}: ${message}`);
+  });
+
+  objectEqualitySubscriber.on('message', (channel, message) => {
+    const subObj: ReqRes = JSON.parse(message);
+    console.log(`listening ${channel}: ${JSON.stringify(subObj)}`);
+    expect(subObj).toEqual(msg);
+  });
 });
 
-afterAll( async () => {
+afterAll(async () => {
   await mockyeah.close();
   publisher.quit();
-  subscriber.quit();
+  subscriber1.quit();
+  subscriber2.quit();
+  objectEqualitySubscriber.quit();
   await redisContainer.remove({ force: true });
-  
   return new Promise((resolve) => setTimeout(() => resolve(), 1000));
 });
-
 
 describe('Process Message', () => {
 
   it('should reject undefined or null message', () => {
-    expect( processMsg({message: undefined, client: publisher, topic: channel}) ).rejects.toThrow();
+    expect(processMsg({ message: undefined, client: publisher, topic: 'rejection' })).rejects.toThrow();
   });
 
   it('should reject undefined or null client', () => {
-    expect( processMsg({message: msg, client: null, topic: channel}) ).rejects.toThrow();
+    expect(processMsg({ message: msg, client: null, topic: 'rejection' })).rejects.toThrow();
   });
 
   it('should reject empty string topic', () => {
-    expect( processMsg({message: msg, client: publisher, topic: ''}) ).rejects.toThrow();
+    expect(processMsg({ message: msg, client: publisher, topic: '' })).rejects.toThrow();
   });
 
-  it('should be the same message object for pub and sub', async () => {
-    await processMsg({ message: msg, client: publisher, topic: channel });
-    expect(await thePromise()).toEqual(msg);
+  it('should publish with no subscriber', async () => {
+    await processMsg({ message: msg, client: publisher, topic: '0 subscriber' })
+      .then((numOfSub) => {
+        expect(numOfSub).toEqual(0);
+      });
+  });
+
+  it('should publish with 2 subscribers', async () => {
+    const topic = '2 subscribers';
+    subscriber1.subscribe(topic);
+    subscriber2.subscribe(topic);
+    await processMsg({ message: msg, client: publisher, topic: topic })
+      .then((numOfSub) => {
+        expect(numOfSub).toEqual(2);
+      });
+  });
+
+  it('should receive same object on subscribed channel', async () => {
+    const topic = 'same obj';
+    objectEqualitySubscriber.subscribe(topic);
+    await processMsg({ message: msg, client: publisher, topic: topic });
   });
 
 });
@@ -87,12 +109,12 @@ describe('Process Message', () => {
 describe('Relay Service', () => {
 
   it('should reject undefined or null url', () => {
-    const f = () => { relayService({targetUrl: undefined, client: publisher, topic: channel}); }; 
+    const f = () => { relayService({targetUrl: undefined, client: publisher, topic: 'rejection'}); }; 
     expect(f).toThrow();
   });
 
   it('should reject undefined or null client', () => {
-    const f = () => { relayService({targetUrl: 'http://localhost:4001', client: null, topic: channel}); };
+    const f = () => { relayService({targetUrl: 'http://localhost:4001', client: null, topic: 'rejection'}); };
     expect(f).toThrow();
   });
 
@@ -138,8 +160,8 @@ describe('Relay Service', () => {
       expect(subObj).toMatchObject({ method: "POST", reqBody: body, statusCode: 204, url: { url: '/pub-sub', query: { abc: '123', def: '456' } } });
     });
 
-    const path: string = '/pub-sub?abc=123&def=456';
-    const body: Object = { id: '1234' };
+    const path = '/pub-sub?abc=123&def=456';
+    const body = { id: '1234' };
     mockyeah.post(path, { status: 204, json: req => req.body });
     await request(relay).post(path)
       .send(body)
@@ -148,16 +170,16 @@ describe('Relay Service', () => {
       .expect(204);
  });
 
-  it('should receive 400 when post text when content-type json', async () => {
+//   it('should receive 400 when post text when content-type json', async () => {
 
-    const path: string = '/pub-sub?abc=123&def=456';
-    const body: string = 'I am a text';
-    mockyeah.post(path, { status: 204, text: req => req.body });
-    await request(relay).post(path)
-      .send(body)
-      .set('Content-Type', 'application/json')
-      .set('Accept', 'application/json')
-      .expect(400);
-  });
+//     const path = '/pub-sub?abc=123&def=456';
+//     const body = 'I am a text';
+//     mockyeah.post(path, { status: 204, text: req => req.body });
+//     await request(relay).post(path)
+//       .send(body)
+//       .set('Content-Type', 'application/json')
+//       .set('Accept', 'application/json')
+//       .expect(400);
+//   });
 
 });
