@@ -5,13 +5,17 @@ import { relayService } from './relayService';
 import retryStrategy from 'node-redis-retry-strategy';
 
 const TARGET_URL = process.env.TARGET_URL;
-const SERVICE_PORT = (process.env.SERVICE_PORT || 80) as number;
+const SERVICE_PORT = (parseInt(process.env.SERVICE_PORT) || 80) as number;
 const REDIS_HOST = process.env.REDIS_HOST;
-const REDIS_PORT = (process.env.REDIS_PORT || 6379) as number;
+const REDIS_PORT = (parseInt(process.env.REDIS_PORT) || 6379) as number;
 const TOPIC = process.env.REDIS_TOPIC;
 
 const logger = getLogger('[relay] app.js');
 const client = redis.createClient({host: REDIS_HOST, port: REDIS_PORT, retry_strategy: retryStrategy });
+
+client.on('error', (err) => {
+  logger.error(`error for ${REDIS_HOST}:${REDIS_PORT} - ${err}`);
+}); 
 
 try {
   relayService({
